requirements.txt
Django==1.6.1
PIL==1.1.7
PyRSS2Gen==1.1
PyYAML==3.10
Twisted==13.2.0
Unidecode==0.04.14
altgraph==0.10.2
bdist-mpkg==0.5.0
biplist==0.6
defusedxml==0.4.1
dj-database-url==0.2.2
dj-static==0.0.5
django-tastypie==0.10.0
django-tastypie-mongoengine==0.4.5
gunicorn==18.0
libcdmi==0.1.2
macholib==1.5.1
modulegraph==0.10.4
mongoengine==0.8.1
numpy==1.8.0
oauthlib==0.6.1
pip-tools==0.3.4
pymongo==2.6.3
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
python-slugify==0.0.7
python-social-auth==0.1.20
pytz==2013.9
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.3.2
six==1.5.2
static==1.0.2
wsgiref==0.1.2
